improve trackGoal route, edit schema, add db query

+++ src/routes/trackGoalRoute.ts
@@ -1,5 +1,6 @@
 import { Env } from '../types';
 import { verifyToken } from '../utils/auth';
+import { checkGoalExists } from '../utils/db_queries';
 
 export const trackGoalRoute = async (request: Request, env: Env): Promise<Response> => {
 	const authResponse = await verifyToken(request, env);
@@ -14,20 +15,34 @@ export const trackGoalRoute = async (request: Request, env: Env): Promise<Respon
 			headers: { 'Content-Type': 'application/json' },
 		});
 	}
-
-	const trackedGoal = await env.DB.prepare(`SELECT * FROM TrackedGoals WHERE user_id = ?`).bind(user.user_id).first();
-	if (trackedGoal) {
-		await env.DB.prepare(`DELETE FROM TrackedGoals WHERE user_id = ?`).bind(user.user_id).run();
-	}
-
-	const { success } = await env.DB.prepare(`INSERT INTO TrackedGoals (goal_id, user_id() VALUES (?, ?)`).bind(goal_id, user.user_id).run();
-	if (success) {
-		return new Response(JSON.stringify({ message: 'User added successfully' }), {
-			status: 200,
+	const goalExists = checkGoalExists(goal_id, env);
+	if (!goalExists) {
+		return new Response(JSON.stringify({ error: 'Goal not found' }), {
+			status: 404,
 			headers: { 'Content-Type': 'application/json' },
 		});
-	} else {
-		return new Response(JSON.stringify({ error: 'Failed to insert user' }), {
+	}
+
+	try {
+		const trackedGoal = await env.DB.prepare(`SELECT * FROM TrackedGoals WHERE user_id = ?`).bind(user.user_id).first();
+		if (trackedGoal) {
+			await env.DB.prepare(`DELETE FROM TrackedGoals WHERE user_id = ?`).bind(user.user_id).run();
+		}
+		const { success } = await env.DB.prepare(`INSERT INTO TrackedGoals (goal_id, user_id) VALUES (?, ?)`).bind(goal_id, user.user_id).run();
+		if (success) {
+			return new Response(JSON.stringify({ message: 'User added successfully' }), {
+				status: 200,
+				headers: { 'Content-Type': 'application/json' },
+			});
+		} else {
+			return new Response(JSON.stringify({ error: 'Failed to insert trackedgoal' }), {
+				status: 500,
+				headers: { 'Content-Type': 'application/json' },
+			});
+		}
+	} catch (error) {
+		console.log('Error tracking goal', error);
+		return new Response(JSON.stringify({ error: 'Failed to insert trackedgoal' }), {
 			status: 500,
 			headers: { 'Content-Type': 'application/json' },
 		});
